+++ arduino-ide-extension/src/browser/arduino-commands.ts
@@ -8,13 +8,6 @@ export namespace ArduinoCommands {
     id: 'arduino-toggle-compile-for-debug',
   };
 
-  /**
-   * Unlike `OPEN_SKETCH`, it opens all files from a sketch folder. (ino, cpp, etc...)
-   */
-  export const OPEN_SKETCH_FILES: Command = {
-    id: 'arduino-open-sketch-files',
-  };
-
   export const OPEN_BOARDS_DIALOG: Command = {
     id: 'arduino-open-boards-dialog',
   };
