+++ arduino-ide-extension/src/browser/theia/core/frontend-application.ts
@@ -4,7 +4,7 @@ import { CommandService } from '@theia/core/lib/common/command';
 import { WorkspaceService } from '@theia/workspace/lib/browser/workspace-service';
 import { FrontendApplication as TheiaFrontendApplication } from '@theia/core/lib/browser/frontend-application';
 import { SketchesService } from '../../../common/protocol';
-import { ArduinoCommands } from '../../arduino-commands';
+import { OpenSketchFiles } from '../../contributions/open-sketch-files';
 
 @injectable()
 export class FrontendApplication extends TheiaFrontendApplication {
@@ -25,7 +25,7 @@ export class FrontendApplication extends TheiaFrontendApplication {
     this.workspaceService.roots.then(async (roots) => {
       for (const root of roots) {
         await this.commandService.executeCommand(
-          ArduinoCommands.OPEN_SKETCH_FILES.id,
+          OpenSketchFiles.Commands.OPEN_SKETCH_FILES.id,
           root.resource
         );
         this.sketchesService.markAsRecentlyOpened(root.resource.toString()); // no await, will get the notification later and rebuild the menu
